simplify Handler class and extract to own file
